requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0

# MongoDB drivers (Compatible versions)
motor==3.1.1
pymongo==4.3.3

# Pydantic & validators
pydantic==1.10.7
email-validator==1.3.1

# Auth & security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Environment loader
python-dotenv==1.0.0
